api/v1alpha1: simplify results API
Right now, the results API requires all inputs, outputs, and trueOutputs
to be stringified representations of their raw values, which in the
common case of JSON inputs and outputs means an unnecessary additional
step of escaping all of the JSON before sending it to the model-tracking
API. This commit simplifies the results API and brings it in line with
the schemas API, allowing complex embedded values to be sent as their
raw type rather than as a string. Notably, this is still safe because
all values for inputs and outputs sent to the results API are checked
against a schema.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>