api/v1alpha1: automatically create versions
This commit adds new capabilities to the API by allowing the server to
automatically create a version for a model if the model declares a
default schema. The semantics of the automatic create allow users to
declare a default schema for a model, which will be used to
automatically create a version for the model if a result is submitted to
the API for a nonexistent version. This is convenient because it allows
users to submit results for models without caring if the version already
exists.

In order to accomplish this, the commit enhances the store and server
code by adding contexts to all of the SQL queries and by wrapping them
in transactions where appropriate.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>